checkpoint

+++ redirect/click-router/src/model/route.rs
@@ -106,7 +106,7 @@ pub struct Route {
     #[serde(default)]
     pub dest_format: DestinationFormat,
     pub code: Option<u16>,
-    pub ttl: Option<u128>,
+    pub ttl: Option<u64>,
     #[serde(default)]
     pub status: RouteStatus,
     #[serde(default)]
